Make dependencies truly optionel + clean up / minimize android build
both Phloc and Schematron was not truly optionel since they was referred
from different files. So the references has been moved to a factory and
ObjectUtil extended for some of the missing methods, where Schematron
was used. Android pom file is cleaned so that all artifcates are build
using the same base configuration and then schematron, phloc and some
server components have been removed from artifact to minimize android
artifacts.