Fix questionnaire schematron with enableWhen and hasAnswer
Without this fix, the 'hasAnswer' element of a questionnaire item is ignored.
This in turn allowes it that one can have both an 'answer' and a 'hasAnswer' item.

A fix is applied for both DSTU3 and R4 and tests are provided.

The tests for R4 are currently ignored, since they require a valid schematron,
e.g. by merging #869