First pass at InMemoryMatcher - all tests passing
Only Token supported atm

Moved code out of SearchBuilder into SearchParameterMap.clean() so it can be used by inMemoryMatcher

Introduced a new composite subscription matcher that tries to match in memory and if it finds a parameter in the criteria it doesn't support, it falls back to the database matcher.

Synchronized Observation lists in test

Fixed a test that wasn't passing due to a changed interface that was mocked.