Update old removed Chef syntax in Vagrant provision
The original reported issue was due to the deprecation and removal of
`Chef::Platform.set`. This was fixed by moving to the new provider
syntax.

Other changes in the Chef cookbooks were performed to address similar
changes due to deprecated and removed methods and models as well as to
conform to new enforcements.

One issue not addressed in this change is to make the Oracle JDK
download work, since Java 8u20 is no longer available.

This fix was tested with the following override in the Vagrantfile:

java: {
  install_flavor: 'openjdk',
  jdk_version: 7
}

The above change is NOT included in this patch set. Further work
would be necessary to update the Oracle download before `vagrant up`
will work without this modification.

Fix #1123

Signed-off-by: Steve Lewis <steven.lewis@cambiahealth.com>