Fixes #3289 Javadoc inverted the boolean value
The method returned the opposite of what the Javadoc declares it will do. I assume that the default implementation is right and the Javadoc contained an error. I think it's somewhat confusing to have Javadoc say A and the method do B. Aligning them to each other will help implementors to understand clearly what is expected.