Fixing concurrency issue in WorkerContextValidationSupportAdapter (#6554) (#6648)
* Fixing concurrency issue in VersionSpecificWorkerContextWrapper (#6554)

When validating if a code is valid against a ValueSet, a potentially
cached (and therefore shared) result object was mutated. This mutation
is now done on a copy of the result object.

* Added test case for concurrent validateCode with issue in VS and CS

The added test case checks if concurrent validation of a code, that
is neither in the ValueSet nor in the CodeSystem creates a validation
result with the CS issue added to the validation result correctly.

* Credit for #6648

---------

Co-authored-by: James Agnew <jamesagnew@gmail.com>